+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.102
+
+### Patch Changes
+
+- Updated dependencies [92244b8]
+    - @sap-ux/odata-service-inquirer@2.16.0
+
+## 0.2.101
+
+### Patch Changes
+
+- Updated dependencies [1d3e398]
+    - @sap-ux/odata-service-inquirer@2.15.0
+
 ## 0.2.100
 
 ### Patch Changes